feat: Add datasets (#45)
Add datasets that will be used as input to cost functions. Also change sphinx config to render init docstrings (and let these take precedence over class docstrings to avoid summary duplication). The reasoning is that if we want to document init parameters, we should do this in the init docstring rather than the class docstring so that ruff is able to detect missing/misspelled parameters in the docstring.