docs: Put __init__ param doc in class doc (#46)
Put doc for __init__ params in class docstring instead of in __init__ docstring to keep both summary and param docs in a single place. This is cleaner and only requires updating one docstring instead of two when making a change. It also allows us to partially document params, i.e. skip self-explanatory ones, unlike in __init__ docstrings where it's all or nothing.