docs(metrics): Add reusable definitions (#81)
Add definitions used in multiple docstrings to snippet files which can
be loaded in docstrings using the include-directive. Using this pattern,
docstrings don't have to redefine defintions, and future changes can be
done in a single place. As part of this, metric definitions have been
moved closer to the default metrics: before, you had to go down 2 levels
- first to data extractors and then to metric utils - to see the
definition of certain metrics, now the definitions are shown directly in
data extractors.