fix(agent): Enable access to custom cost function methods
Remove cost function proxy so that the original cost function can be accessed from algorithms. This is necessary for algorithms that rely on more methods than CostFunction provides. This is a work around, in the future we should add more cost function abstracts to not have algorithms depend on methods that may or may not exist.