feat(decent-bench): Generalize optimization parameter (#223)
Implements a new array type called `Array`, alongside some helper
typealiases/enums for `SupportedArrayTypes`, `SupportedFrameworks` and
`SupportedDevices`.

* This new `Array` class make operations between different frameworks
agnostic alongside the interoparability layer. No automatic casting
between frameworks are performed due to performance. However, this
should not be an issue as the only reason two different frameworks would
be used is if a dataset is used with an incompatible Cost.
* During typechecking the new `Array` class is used throughout Decent
Bench while during runtime the `Array` class is unwrapped to its base
framework for performance reasons.
* Modifications have been made to the interoperability layer so that it
operates on `Array`. Some more functionality has been added and better
modularization.
* A custom decorator for cost functions is added to allow cost functions
to operate directly on its underlying framework while allowing for type
checking. During runtime, this decorator will unpack the `Array` object
and convert it to the desired framework silently if needed. If
conversion is needed it will throw a warning to tell the user that this
behavior might cause performance issues.
* Datasets have been updated to return `tuple[Array, Array]`.
* Network weights updated to use the new `Array` class, a new adjacency
property which returns the adjacency matrix and a set_weights function
is added which all should solve issue #203.
* Documentation updates to reflect these changes. 
* Minor fixes for issues created by updating sphinx 8.2.3 -> 9.0.1.
* Optional development environment with gpu support for tensorflow/jax. 

closes #200, #203, #224